+++ pkg/utils/utils_pkg.go
@@ -40,8 +40,8 @@ func FromNullTimeRFC3339(nt any) string    { return utils.FromNullTimeRFC3339(nt
 func ToNullInt16(i int16) sql.NullInt16    { return utils.ToNullInt16(i) }
 func FromNullInt16(ni sql.NullInt16) int16 { return utils.FromNullInt16(ni) }
 
-func ToNullInt64(i int) sql.NullInt64    { return utils.ToNullInt64(i) }
-func FromNullInt64(ni sql.NullInt64) int { return utils.FromNullInt64(ni) }
+func ToNullInt64(i int64) sql.NullInt64    { return utils.ToNullInt64(i) }
+func FromNullInt64(ni sql.NullInt64) int64 { return utils.FromNullInt64(ni) }
 
 func ToNullBool(b bool) sql.NullBool    { return utils.ToNullBool(b) }
 func FromNullBool(nb sql.NullBool) bool { return utils.FromNullBool(nb) }
